control_podium/requirements.txt
pyserial
pyyaml
bleak>=0.21
PySide6>=6.6
